backend_module/requirements.txt
tensorflow==2.16.1
pandas==2.0.3
scikit-learn==1.5.0
numpy==1.24.3
joblib==1.3.2
Django==5.0.7
django-cors-headers==4.4.0
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
pip~=25.0.1
wheel~=0.45.1
gast~=0.6.0
h5py~=3.13.0
pytz~=2025.2
typing_extensions~=4.13.1
rich~=14.0.0
Pygments~=2.19.1
keras~=3.9.2
optree~=0.14.1
packaging~=24.2
scipy~=1.15.2
requests~=2.32.3
namex~=0.0.8
mdurl~=0.1.2
setuptools~=78.1.0
six~=1.17.0
threadpoolctl~=3.6.0
asgiref~=3.8.1
sqlparse~=0.5.3
MarkupSafe~=3.0.2
python-dateutil~=2.9.0.post0
certifi~=2025.1.31
urllib3~=2.3.0
Markdown~=3.7
idna~=3.10
termcolor~=3.0.1
opt_einsum~=3.4.0
flatbuffers~=25.2.10
wrapt~=1.17.2
astunparse~=1.6.3
tensorboard~=2.16.2
Werkzeug~=3.1.3
transformers==4.30.0
torch==2.0.1
django-environ==0.10.0
filelock~=3.18.0
colorama~=0.4.6
PyYAML~=6.0.2
huggingface-hub~=0.30.1
regex~=2024.11.6
mpmath~=1.3.0
sympy~=1.13.1
tqdm~=4.67.1
Jinja2~=3.1.6
fsspec~=2025.3.2
networkx~=3.4.2
tokenizers~=0.13.3
safetensors~=0.5.3
schedule==1.2.0